+++ Incorpreal/Battle/Battle.cs
@@ -10,15 +10,12 @@ public class Battle : Node
     public GlobalPlayer gp;
     public int currentFighterPos;
     public Label playerHP;
-    
-    
+
     public Battle()
     {
 
     }
-    public AudioStreamPlayer audioStreamPlayer = new AudioStreamPlayer();
-    AudioStream HammerAttack = (AudioStream)GD.Load("res://sounds/HammerClank.wav");
-    AudioStream BattleStart = (AudioStream)GD.Load("res://sounds/scaletest.wav");
+
     public override void _Ready()
     {
         gp = (GlobalPlayer)GetNode("/root/GlobalData");
@@ -33,9 +30,6 @@ public class Battle : Node
         btn = GetNode<Button>("Button2") as Button;
         gp.updateHealthLabel(playerHP);
         updateEnemyHealth();
-        this.AddChild(audioStreamPlayer);
-        audioStreamPlayer.Stream = BattleStart;
-        audioStreamPlayer.Play();
     }
 
     //This is just for demo expamples to not get stuck on battle screen.
@@ -46,8 +40,6 @@ public class Battle : Node
 
     public void _on_TakeDamageButton_pressed()
     {
-        audioStreamPlayer.Stream = HammerAttack;
-        audioStreamPlayer.Play();
         GlobalPlayer gp = (GlobalPlayer)GetNode("/root/GlobalData");
         gp.takeDamage(5);
         var healthLabel = GetNode<Label>("HealthLabel") as Label;
