+++ Incorpreal/Enemies/Bat.tscn
@@ -5,16 +5,12 @@
 [sub_resource type="CircleShape2D" id=1]
 radius = 18.3538
 
-[node name="Bat" type="KinematicBody2D" groups=[
-"Enemies",
-]]
-collision_layer = 16
-collision_mask = 7
+[node name="Bat" type="KinematicBody2D"]
 
 [node name="CollisionShape2D" type="CollisionShape2D" parent="."]
 position = Vector2( 0, -5 )
 shape = SubResource( 1 )
 
-[node name="Sprite" type="Sprite" parent="."]
+[node name="bat" type="Sprite" parent="."]
 position = Vector2( 2, 2 )
 texture = ExtResource( 1 )
