+++ Incorpreal/Enemies/Bear.tscn
@@ -5,11 +5,9 @@
 [sub_resource type="CircleShape2D" id=1]
 radius = 39.3799
 
-[node name="Bear" type="KinematicBody2D" groups=[
-"Enemies",
-]]
+[node name="Bear" type="KinematicBody2D"]
 
-[node name="Sprite" type="Sprite" parent="."]
+[node name="bear" type="Sprite" parent="."]
 texture = ExtResource( 1 )
 
 [node name="CollisionShape2D" type="CollisionShape2D" parent="."]
