+++ Incorpreal/Game.tscn
@@ -1,23 +1,33 @@
-[gd_scene load_steps=11 format=2]
+[gd_scene load_steps=14 format=2]
 
-[ext_resource path="res://Walls/Low Wall.tscn" type="PackedScene" id=1]
-[ext_resource path="res://Walls/Ghost Wall.tscn" type="PackedScene" id=2]
-[ext_resource path="res://Enemies/Bat.tscn" type="PackedScene" id=3]
+[ext_resource path="res://assets/wall.png" type="Texture" id=1]
+[ext_resource path="res://assets/ghostwall.png" type="Texture" id=2]
+[ext_resource path="res://assets/player.png" type="Texture" id=3]
 [ext_resource path="res://bat/Bat.cs" type="Script" id=4]
-[ext_resource path="res://Walls/Wall.tscn" type="PackedScene" id=5]
+[ext_resource path="res://assets/lowwall.png" type="Texture" id=5]
 [ext_resource path="res://assets/bat.png" type="Texture" id=6]
 [ext_resource path="res://Game.cs" type="Script" id=7]
-[ext_resource path="res://Enemies/Bear.tscn" type="PackedScene" id=9]
-[ext_resource path="res://player/Player.tscn" type="PackedScene" id=10]
+[ext_resource path="res://player/Player.cs" type="Script" id=8]
 
 [sub_resource type="RectangleShape2D" id=1]
+extents = Vector2( 33, 33 )
+
+[sub_resource type="RectangleShape2D" id=2]
+extents = Vector2( 33.125, 32.75 )
+
+[sub_resource type="RectangleShape2D" id=3]
 extents = Vector2( 32, 31 )
+
+[sub_resource type="RectangleShape2D" id=4]
+extents = Vector2( 31.919, 32.0566 )
+
+[sub_resource type="RectangleShape2D" id=5]
+extents = Vector2( 31.8679, 32 )
 
 [node name="Game" type="Node"]
 script = ExtResource( 7 )
 
 [node name="Map" type="Node2D" parent="."]
-
 
 [node name="Ghost Wall" type="StaticBody2D" parent="Map"]
 collision_layer = 4
@@ -33,24 +43,24 @@ texture = ExtResource( 2 )
 position = Vector2( 110.037, 68.0034 )
 shape = SubResource( 1 )
 
-[node name="Bear" parent="." instance=ExtResource( 9 )]
-position = Vector2( 111.118, 207.95 )
+[node name="Wall" type="StaticBody2D" parent="Map"]
 
-[node name="Player" parent="." instance=ExtResource( 10 )]
-position = Vector2( 316.269, 217.27 )
+[node name="Sprite" type="Sprite" parent="Map/Wall"]
 
-[node name="Ghost Wall" parent="." instance=ExtResource( 2 )]
-position = Vector2( 122.523, 109.101 )
+[node name="wall" type="Sprite" parent="Map/Wall/Sprite"]
+position = Vector2( 292, 69 )
+texture = ExtResource( 1 )
 
-[node name="Bat2" parent="." instance=ExtResource( 3 )]
-position = Vector2( 435.756, 217.633 )
+[node name="CollisionShape2D" type="CollisionShape2D" parent="Map/Wall"]
+position = Vector2( 292, 69 )
+shape = SubResource( 2 )
 
-[node name="Bat" type="KinematicBody2D" parent="."]
+[node name="Bat" type="KinematicBody2D" parent="Map"]
 position = Vector2( 93, 67 )
 collision_layer = 16
 script = ExtResource( 4 )
 
-[node name="Sprite" type="Sprite" parent="Bat"]
+[node name="Sprite" type="Sprite" parent="Map/Bat"]
 
 [node name="bat" type="Sprite" parent="Map/Bat/Sprite"]
 position = Vector2( 387.919, 126.454 )
@@ -98,19 +108,9 @@ script = ExtResource( 8 )
 position = Vector2( 267, 197 )
 texture = ExtResource( 3 )
 
-[node name="bat" type="Sprite" parent="Bat/Sprite"]
-position = Vector2( 425, 154 )
-texture = ExtResource( 6 )
-
-[node name="CollisionShape2D" type="CollisionShape2D" parent="Bat" groups=[
-"Enemies",
-]]
-position = Vector2( 426, 152.159 )
-shape = SubResource( 1 )
-
-[node name="Wall" parent="." instance=ExtResource( 5 )]
-position = Vector2( 314.127, 103.11 )
-
+[node name="CollisionShape2D" type="CollisionShape2D" parent="Map/Player"]
+position = Vector2( 267, 196.941 )
+shape = SubResource( 5 )
 
 [node name="HealthLabel" type="Label" parent="Map"]
 margin_left = 324.288
@@ -121,7 +121,3 @@ text = "Health Goes Here"
 __meta__ = {
 "_edit_use_anchors_": false
 }
-
-[node name="Low Wall" parent="." instance=ExtResource( 1 )]
-position = Vector2( 504.335, 114.625 )
-
