+++ Incorpreal/Gem/Gem.tscn
@@ -10,28 +10,27 @@ resource_name = "Bounce"
 length = 5.0
 loop = true
 tracks/0/type = "value"
-tracks/0/path = NodePath("gem:offset")
+tracks/0/path = NodePath(".:position")
 tracks/0/interp = 1
 tracks/0/loop_wrap = true
 tracks/0/imported = false
 tracks/0/enabled = true
 tracks/0/keys = {
 "times": PoolRealArray( 0, 1, 2 ),
-"transitions": PoolRealArray( 0.420448, 0.68302, 0.68 ),
+"transitions": PoolRealArray( 0.420448, 0.420448, 1 ),
 "update": 0,
 "values": [ Vector2( 0, 0 ), Vector2( 0, -3 ), Vector2( 0, -0.25 ) ]
 }
 
 [node name="Gem" type="Area2D"]
+position = Vector2( 0, -0.0600272 )
 
 [node name="gem" type="Sprite" parent="."]
 texture = ExtResource( 1 )
-offset = Vector2( 0, -0.00715218 )
 
 [node name="CollisionShape2D" type="CollisionShape2D" parent="."]
 position = Vector2( -1, -3 )
 shape = SubResource( 1 )
 
 [node name="AnimationPlayer" type="AnimationPlayer" parent="."]
-autoplay = "Bounce"
 anims/Bounce = SubResource( 2 )
