+++ Incorpreal/Walls/Ghost Wall.tscn
@@ -10,7 +10,7 @@ extents = Vector2( 31.4453, 30.5435 )
 ]]
 position = Vector2( 284.523, 160.101 )
 collision_layer = 4
-collision_mask = 16
+collision_mask = 0
 
 [node name="Sprite" type="Sprite" parent="."]
