+++ Incorpreal/player/Player.cs
@@ -5,14 +5,9 @@ public class Player : KinematicBody2D {
     [Export]
 
     public int moveSpeed = 125;
-    public PhysicsBody2D possessee = null;
-    public string resPath;
-    public Map map = new Map();
     public CollisionShape2D hitbox;
-    public Sprite playerSpriteNode;
     public AnimationPlayer animate;
     public GlobalPlayer gp;
-
 
     //For all the methods pertaining to stats, nothing is set in stone
     //numbers are expected to change as at a later date.
@@ -114,12 +109,7 @@ public class Player : KinematicBody2D {
                 //ChangeState("dead");
                 collision.Collider.Call("Hit");
             }
-        } 
-
-        //Possession listener
-        if (Input.IsActionJustPressed("possession")) { //If R is pressed
-            Possess();
-        }       
+        }        
     }
 
     // Imported from Elijah's branch, and matched names, and styles
@@ -146,51 +136,4 @@ public class Player : KinematicBody2D {
 	  }
 	}
     */
-
-     public void Possess() {
-        //1. Check if anyone within range
-        Area2D possessionArea = (Area2D)GetNode("Area2D");
-        Godot.Collections.Array nearby = possessionArea.GetOverlappingBodies(); //Check who is nearby
-            float closestDistance = 1000;
-            int closestEnemyIndex = 0;
-            Boolean enemyFound = false;
-
-            //2. Find closest enemy
-            for (int x = 0; x < nearby.Count; x++) { //Iterate them
-                PhysicsBody2D currentEnemy = (PhysicsBody2D) nearby[x]; //Grab one
-                if (currentEnemy.GetGroups().Contains("Enemies")) { //Skip bodies not belonging to the Enemies group
-                    float currentDistance = currentEnemy.GlobalPosition.DistanceTo(this.GlobalPosition); //Calculate distance
-                    if (currentDistance < closestDistance) { //Check if closer than current closest
-                        closestEnemyIndex = x;
-                        closestDistance = currentDistance;
-                        enemyFound = true;
-                    }
-                }
-            }  
-            
-            //3. If suitable enemy found & player not already possessing someone, possess that enemy
-            if (enemyFound && this.possessee == null) {
-                possessee = (KinematicBody2D) nearby[closestEnemyIndex]; //Grab victim
-                this.resPath = possessee.Filename; //Grab victim resource path for later
-                Sprite victimSprite = (Sprite) possessee.GetNode("Sprite"); //Grab victim sprite
-                this.SetCollisionMaskBit(2, true); //Make GhostWalls impenetrable while possessing
-                if (resPath.Contains("Bat")) {
-                    this.SetCollisionLayerBit(0, false); //If possessing a bat, gain ability to fly over LowWalls. This was the only way it worked...
-                    this.SetCollisionMaskBit(3, false); //Turn off LowWall collisions
-                }
-                //Possession animation here (optional)
-                this.playerSpriteNode = (Sprite) GetNode("Sprite/player");
-                playerSpriteNode.Texture = victimSprite.Texture; //Copy victim's texture
-                victimSprite.GetParent().QueueFree(); //Make enemy disappear
-            } else if (possessee != null) { //Else if already possessing, undo it
-                playerSpriteNode.Texture = (Texture) ResourceLoader.Load("res://assets/player.png"); //Return player sprite to normal
-                this.SetCollisionMaskBit(2, false); //Make GhostWalls penetrable again
-                if (resPath.Contains("Bat")) { //Return from Bat mode
-                    this.SetCollisionLayerBit(0, true);
-                    this.SetCollisionMaskBit(3, true);
-                }
-                this.map.SpawnEnemy(this.resPath, this.GlobalPosition, GetTree().CurrentScene); //Bring original enemy back
-                possessee = null;
-            }
-    }
 }
