+++ Incorpreal/sounds/scale test.wav.import
@@ -1,0 +1,21 @@
+[remap]
+
+importer="wav"
+type="AudioStreamSample"
+path="res://.import/scale test.wav-300c2548eb3e63295d14ea1343dd44b6.sample"
+
+[deps]
+
+source_file="res://sounds/scale test.wav"
+dest_files=[ "res://.import/scale test.wav-300c2548eb3e63295d14ea1343dd44b6.sample" ]
+
+[params]
+
+force/8_bit=false
+force/mono=false
+force/max_rate=false
+force/max_rate_hz=44100
+edit/trim=false
+edit/normalize=false
+edit/loop=false
+compress/mode=1
